Rename index.js to msstreams-utils.js
