+++ msstreams-utils.js
